+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.1.5 - Unreleased
 
+### Added
+* UnknownBackendError and DuplicateJobError
+
 
 ## Version 0.1.4 - Released 2022-10-31